backend/pyproject.toml
[project]
name = "pm-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "python-dotenv>=1.1",
    "pyjwt>=2.10",
    "openai>=1.82",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "httpx>=0.28",
]
